titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/airpay-config/data-access-airpay-config/src/lib/+state/airpay-configuration.selectors.ts
import { createSelector } from '@ngrx/store';
import { airpayConfigurationState } from './airpay-configuration.reducer';

const selectVendorList = createSelector(
  airpayConfigurationState,
  state => state.vendorList
);

const selectTotalElements = createSelector(
  airpayConfigurationState,
  state => state.totalCount
);

const selectHasError = createSelector(
  airpayConfigurationState,
  state => state.hasError
);

const selectIsLoading = createSelector(
  airpayConfigurationState,
  state => state.isLoading
);

export const AirpayConfigurationSelectors = {
  selectHasError,
  selectIsLoading,
  selectVendorList,
  selectTotalElements
};
